Falta implementar bien object sistemaventas y hacer los tests

+++ VentasTpAnual/src/TipoCliente.scala
@@ -1,4 +1,4 @@
 trait TipoCliente {
-		def dtoTipoPers(precioBase:Float):Double
+		def dtoTipoPers(precioBase:Double):Double
 		def datosLegales():String
 }
